+++ asn1_parser/per_opentype.c
@@ -15,7 +15,7 @@ typedef struct uper_ugot_key {
 } uper_ugot_key;
 
 static int uper_ugot_refill(asn_per_data_t *pd);
-static int per_skip_bits(asn_per_data_t *pd, int skip_nbits);
+static int per_skip_bits(asn_per_data_t *pd, ssize_t skip_nbits);
 static asn_dec_rval_t uper_sot_suck(asn_codec_ctx_t *, asn_TYPE_descriptor_t *td, asn_per_constraints_t *constraints, void **sptr, asn_per_data_t *pd);
 
 /*
@@ -356,10 +356,10 @@ uper_ugot_refill(asn_per_data_t *pd) {
 }
 
 static int
-per_skip_bits(asn_per_data_t *pd, int skip_nbits) {
+per_skip_bits(asn_per_data_t *pd, ssize_t skip_nbits) {
 	int hasNonZeroBits = 0;
 	while(skip_nbits > 0) {
-		int skip;
+		ssize_t skip;
 
 		/* per_get_few_bits() is more efficient when nbits <= 24 */
 		if(skip_nbits < 24)
